backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.2
redis==5.1.1
httpx==0.27.2
alembic==1.13.3
PyJWT[crypto]>=2.8.0
bcrypt>=4.1.0
cryptography==43.0.3
slowapi==0.1.9
gunicorn==23.0.0
email-validator==2.2.0
python-multipart==0.0.12
